feat(validator): add PR reviews
