+++ kroma-validator/cmd/validator/flags.go
@@ -1,4 +1,4 @@
-package main
+package validator
 
 import (
 	"github.com/urfave/cli/v2"
@@ -6,7 +6,7 @@ import (
 
 var TokenAmountFlag = &cli.StringFlag{
 	Name:     "amount",
-	Usage:    "Amount of governance token (in wei)",
+	Usage:    "Amount of asset token (in wei)",
 	Required: true,
 }
 
@@ -18,12 +18,12 @@ var EthAmountFlag = &cli.StringFlag{
 
 var CommissionRateFlag = &cli.Uint64Flag{
 	Name:     "commission-rate",
-	Usage:    "The commission rate the validator sets (in percentage). Maximum 100.",
+	Usage:    "The commission rate earned by the validator (in percentage). Maximum 100.",
 	Required: true,
 }
 
 var CommissionMaxChangeRateFlag = &cli.Uint64Flag{
 	Name:     "commission-max-change-rate",
-	Usage:    "The maximum changeable commission rate change (in percentage). Maximum 100.",
+	Usage:    "The maximum changeable commission rate at once (in percentage). Maximum 100.",
 	Required: true,
 }
